+++ src/lib/group.ts
@@ -1,4 +1,4 @@
-import { Field, FieldVar, isField } from './field.js';
+import { Field, FieldVar } from './field.js';
 import { Scalar } from './scalar.js';
 import { Snarky } from '../snarky.js';
 import { Field as Fp } from '../provable/field-bigint.js';
@@ -48,10 +48,10 @@ class Group {
     x: FieldVar | Field | number | string | bigint;
     y: FieldVar | Field | number | string | bigint;
   }) {
-    this.x = isField(x) ? x : new Field(x);
-    this.y = isField(y) ? y : new Field(y);
+    this.x = x instanceof Field ? x : new Field(x);
+    this.y = y instanceof Field ? y : new Field(y);
 
-    if (this.#isConstant()) {
+    if (isConstant(this)) {
       // we also check the zero element (0, 0) here
       if (this.x.equals(0).and(this.y.equals(0)).toBoolean()) return;
 
@@ -72,31 +72,6 @@ class Group {
     }
   }
 
-  // helpers
-  static #fromAffine({ x, y, infinity }: GroupAffine) {
-    return infinity ? Group.zero : new Group({ x, y });
-  }
-
-  static #fromProjective({ x, y, z }: { x: bigint; y: bigint; z: bigint }) {
-    return this.#fromAffine(Pallas.toAffine({ x, y, z }));
-  }
-
-  #toTuple(): [0, FieldVar, FieldVar] {
-    return [0, this.x.value, this.y.value];
-  }
-
-  #isConstant() {
-    return this.x.isConstant() && this.y.isConstant();
-  }
-
-  #toProjective() {
-    return Pallas.fromAffine({
-      x: this.x.toBigInt(),
-      y: this.y.toBigInt(),
-      infinity: false,
-    });
-  }
-
   /**
    * Checks if this element is the `zero` element `{x: 0, y: 0}`.
    */
@@ -114,15 +89,15 @@ class Group {
    * ```
    */
   add(g: Group) {
-    if (this.#isConstant() && g.#isConstant()) {
+    if (isConstant(this) && isConstant(g)) {
       // we check if either operand is zero, because adding zero to g just results in g (and vise versa)
       if (this.isZero().toBoolean()) {
         return g;
       } else if (g.isZero().toBoolean()) {
         return this;
       } else {
-        let g_proj = Pallas.add(this.#toProjective(), g.#toProjective());
-        return Group.#fromProjective(g_proj);
+        let g_proj = Pallas.add(toProjective(this), toProjective(g));
+        return fromProjective(g_proj);
       }
     } else {
       const { x: x1, y: y1 } = this;
@@ -163,9 +138,9 @@ class Group {
       });
 
       let [, x, y] = Snarky.gates.ecAdd(
-        Group.from(x1.seal(), y1.seal()).#toTuple(),
-        Group.from(x2.seal(), y2.seal()).#toTuple(),
-        Group.from(x3, y3).#toTuple(),
+        toTuple(Group.from(x1.seal(), y1.seal())),
+        toTuple(Group.from(x2.seal(), y2.seal())),
+        toTuple(Group.from(x3, y3)),
         inf.toField().value,
         same_x.value,
         s.value,
@@ -216,13 +191,13 @@ class Group {
   scale(s: Scalar | number | bigint) {
     let scalar = Scalar.from(s);
 
-    if (this.#isConstant() && scalar.isConstant()) {
-      let g_proj = Pallas.scale(this.#toProjective(), scalar.toBigInt());
-      return Group.#fromProjective(g_proj);
+    if (isConstant(this) && scalar.isConstant()) {
+      let g_proj = Pallas.scale(toProjective(this), scalar.toBigInt());
+      return fromProjective(g_proj);
     } else {
       let [, ...bits] = scalar.value;
       bits.reverse();
-      let [, x, y] = Snarky.group.scale(this.#toTuple(), [0, ...bits]);
+      let [, x, y] = Snarky.group.scale(toTuple(this), [0, ...bits]);
       return new Group({ x, y });
     }
   }
@@ -447,4 +422,30 @@ class Group {
       }`;
     }
   }
+}
+
+// internal helpers
+
+function isConstant(g: Group) {
+  return g.x.isConstant() && g.y.isConstant();
+}
+
+function toTuple(g: Group): [0, FieldVar, FieldVar] {
+  return [0, g.x.value, g.y.value];
+}
+
+function toProjective(g: Group) {
+  return Pallas.fromAffine({
+    x: g.x.toBigInt(),
+    y: g.y.toBigInt(),
+    infinity: false,
+  });
+}
+
+function fromProjective({ x, y, z }: { x: bigint; y: bigint; z: bigint }) {
+  return fromAffine(Pallas.toAffine({ x, y, z }));
+}
+
+function fromAffine({ x, y, infinity }: GroupAffine) {
+  return infinity ? Group.zero : new Group({ x, y });
 }